+++ lib/kitchen/directions/bake_unnumbered_tables.rb
@@ -9,6 +9,7 @@ module Kitchen
           table.remove_attribute('summary')
           table.parent.add_class('os-unstyled-container') if table.unstyled?
           table.parent.add_class('os-column-header-container') if table.column_header?
+          table.parent.add_class('os-top-titled-container') if table.top_titled?
         end
       end
     end